Move tests to root module
